requirements.txt
django==2.0.7
django-configurations==2.0
psycopg2-binary==2.7.5
factory-boy==2.11.1
